+++ src/utils/components/DiskModal/DiskFormFields/utils/modesMapping.ts
