--- src/utils/components/DiskModal/index.ts
@@ -1,1 +1,0 @@
-export { default as DiskModal } from '../../../views/virtualmachines/details/tabs/disk/modal/DiskModal';
